moving methods from proof_panel to the new ProofStepView after the merge
